Fix: 그룹 참여 시 예수금 충전 기능 수정

+++ vote-service/src/main/java/com/example/vote_service/service/ProposalService.java
@@ -227,7 +227,7 @@ public class ProposalService {
                     proposal.getGroupId(),
                     UUID.fromString((String) payloadMap.get("stockId")),
                     TradingAction.valueOf(proposal.getAction().name()),
-                    (Integer) payloadMap.get("quantity"),
+                    (Float) payloadMap.get("quantity"),
                     new BigDecimal(payloadMap.get("price").toString()),
                     proposal.getPayload(),
                     null, null, null, null, null, null, null
